streamline cached taxonomy
