+++ tests/test_taxonomy.py
@@ -10,216 +10,12 @@ import pytest
 from esp_data.backends import PandasBackend
 from esp_data.discover import AddTaxonomy, AddTaxonomyConfig, GBIFConverter
 
-
-def _write_gbif_tsv(tmp_path: Path, rows: List[Dict[str, Any]]) -> str:
-    """
-    Write a minimal GBIF-like TSV for testing and return its filepath.
-
-    The GBIFConverter implementation expects at least these columns:
-    - taxonID
-    - canonicalName
-    - taxonomicStatus
-    - taxonRank
-    - parentNameUsageID
-    - acceptedNameUsageID
-    """
-    df = pd.DataFrame(rows)
-
-    # Ensure column presence and ordering (useful for debugging)
-    required_cols = [
-        "taxonID",
-        "canonicalName",
-        "taxonomicStatus",
-        "taxonRank",
-        "parentNameUsageID",
-        "acceptedNameUsageID",
-    ]
-    missing = [c for c in required_cols if c not in df.columns]
-    if missing:
-        raise AssertionError(f"Missing required columns for fixture TSV: {missing}")
-
-    fp = tmp_path / "gbif_animals_minimal.tsv"
-    df.to_csv(fp, sep="\t", index=False)
-    return str(fp)
-
-
-# GBIFConverter Unit Tests
-
-def test_get_accepted_species_info_no_match(tmp_path: Path) -> None:
-    fp = _write_gbif_tsv(
-        tmp_path,
-        rows=[
-            {
-                "taxonID": 1,
-                "canonicalName": "Corvus corax",
-                "taxonomicStatus": "accepted",
-                "taxonRank": "species",
-                "parentNameUsageID": np.nan,
-                "acceptedNameUsageID": np.nan,
-            }
-        ],
-    )
-    converter = GBIFConverter(gbif_animals_tsv_fp=fp, use_precomputed_outputs=False, cache_path=None)
-
-    out, ok = converter("Does not exist")
-    assert out == {}
-    assert ok is False
-
-
-def test_get_accepted_species_info_accepts_species(tmp_path: Path) -> None:
-    fp = _write_gbif_tsv(
-        tmp_path,
-        rows=[
-            {
-                "taxonID": 10,
-                "canonicalName": "Corvus corax",
-                "taxonomicStatus": "accepted",
-                "taxonRank": "species",
-                "parentNameUsageID": np.nan,
-                "acceptedNameUsageID": np.nan,
-            }
-        ],
-    )
-    converter = GBIFConverter(gbif_animals_tsv_fp=fp, use_precomputed_outputs = False, cache_path=None)
-
-    out, ok = converter("Corvus corax")
-    assert ok is True
-    assert out["taxonID"] == 10
-    assert out["taxonomicStatus"] == "accepted"
-    assert out["taxonRank"] == "species"
-    assert out["canonicalName"] == "Corvus corax"
-
-
-def test_get_accepted_species_info_resolves_synonym_to_accepted(tmp_path: Path) -> None:
-    fp = _write_gbif_tsv(
-        tmp_path,
-        rows=[
-            # accepted usage
-            {
-                "taxonID": 100,
-                "canonicalName": "Puma concolor",
-                "taxonomicStatus": "accepted",
-                "taxonRank": "species",
-                "parentNameUsageID": np.nan,
-                "acceptedNameUsageID": np.nan,
-            },
-            # synonym that points to accepted usage
-            {
-                "taxonID": 101,
-                "canonicalName": "Felis concolor",
-                "taxonomicStatus": "synonym",
-                "taxonRank": "species",
-                "parentNameUsageID": np.nan,
-                "acceptedNameUsageID": 100.0,
-            },
-        ],
-    )
-    converter = GBIFConverter(gbif_animals_tsv_fp=fp, use_precomputed_outputs=False, cache_path=None)
-
-    out, ok = converter("Felis concolor")
-    assert ok is True
-    assert out["taxonID"] == 100
-    assert out["canonicalName"] == "Puma concolor"
-    assert out["taxonomicStatus"] == "accepted"
-    assert out["taxonRank"] == "species"
-
-
-def test_get_accepted_species_info_walks_up_from_lower_rank(tmp_path: Path) -> None:
-    fp = _write_gbif_tsv(
-        tmp_path,
-        rows=[
-            # accepted species
-            {
-                "taxonID": 200,
-                "canonicalName": "Canis lupus",
-                "taxonomicStatus": "accepted",
-                "taxonRank": "species",
-                "parentNameUsageID": np.nan,
-                "acceptedNameUsageID": np.nan,
-            },
-            # subspecies that points to parent species
-            {
-                "taxonID": 201,
-                "canonicalName": "Canis lupus familiaris",
-                "taxonomicStatus": "accepted",
-                "taxonRank": "subspecies",
-                "parentNameUsageID": 200.0,
-                "acceptedNameUsageID": np.nan,
-            },
-        ],
-    )
-    converter = GBIFConverter(gbif_animals_tsv_fp=fp, use_precomputed_outputs=False, cache_path=None)
-
-    out, ok = converter("Canis lupus familiaris")
-    assert ok is True
-    assert out["taxonID"] == 200
-    assert out["canonicalName"] == "Canis lupus"
-    assert out["taxonRank"] == "species"
-    assert out["taxonomicStatus"] == "accepted"
-
-
-def test_get_accepted_species_info_duplicate_canonical_prefers_accepted(tmp_path: Path) -> None:
-    fp = _write_gbif_tsv(
-        tmp_path,
-        rows=[
-            # accepted row for same canonicalName
-            {
-                "taxonID": 300,
-                "canonicalName": "Dup name",
-                "taxonomicStatus": "accepted",
-                "taxonRank": "species",
-                "parentNameUsageID": np.nan,
-                "acceptedNameUsageID": np.nan,
-            },
-            # non-accepted duplicate
-            {
-                "taxonID": 301,
-                "canonicalName": "Dup name",
-                "taxonomicStatus": "synonym",
-                "taxonRank": "species",
-                "parentNameUsageID": np.nan,
-                "acceptedNameUsageID": 300.0,
-            },
-        ],
-    )
-    converter = GBIFConverter(gbif_animals_tsv_fp=fp, use_precomputed_outputs=False, cache_path=None)
-
-    out, ok = converter("Dup name")
-    assert ok is True
-    assert out["taxonID"] == 300
-    assert out["taxonomicStatus"] == "accepted"
-    assert out["taxonRank"] == "species"
-    assert out["canonicalName"] == "Dup name"
-
-
-def test_get_accepted_species_info_cycle_is_detected(tmp_path: Path) -> None:
-    fp = _write_gbif_tsv(
-        tmp_path,
-        rows=[
-            # non-species record whose parent points to itself (cycle)
-            {
-                "taxonID": 400,
-                "canonicalName": "Cycle name",
-                "taxonomicStatus": "accepted",
-                "taxonRank": "subspecies",
-                "parentNameUsageID": 400.0,
-                "acceptedNameUsageID": np.nan,
-            }
-        ],
-    )
-    converter = GBIFConverter(gbif_animals_tsv_fp=fp, use_precomputed_outputs=False, cache_path=None)
-
-    out, ok = converter("Cycle name")
-    assert out == {}
-    assert ok is False
-
-
 # AddTaxonomy Transform Unit Tests
 
-def _create_gbif_tsv_with_taxonomy(tmp_path: Path) -> str:
-    """Create a GBIF TSV with full taxonomy info for testing AddTaxonomy."""
-    rows = [
-        {
+def _create_gbif_json_with_taxonomy(tmp_path: Path) -> str:
+    """Create a GBIF json with full taxonomy info for testing AddTaxonomy."""
+    lookupdict = {
+        "Corvus corax": {
             "taxonID": 1,
             "canonicalName": "Corvus corax",
             "taxonomicStatus": "accepted",
@@ -233,7 +29,7 @@ def _create_gbif_tsv_with_taxonomy(tmp_path: Path) -> str:
             "family": "Corvidae",
             "genus": "Corvus",
         },
-        {
+        "Passer domesticus": {
             "taxonID": 2,
             "canonicalName": "Passer domesticus",
             "taxonomicStatus": "accepted",
@@ -247,7 +43,7 @@ def _create_gbif_tsv_with_taxonomy(tmp_path: Path) -> str:
             "family": "Passeridae",
             "genus": "Passer",
         },
-        {
+        "Canis lupus" : {
             "taxonID": 3,
             "canonicalName": "Canis lupus",
             "taxonomicStatus": "accepted",
@@ -261,22 +57,7 @@ def _create_gbif_tsv_with_taxonomy(tmp_path: Path) -> str:
             "family": "Canidae",
             "genus": "Canis",
         },
-        # add one synonym entry
-        {
-            "taxonID": 4,
-            "canonicalName": "Felis concolor",
-            "taxonomicStatus": "synonym",
-            "taxonRank": "species",
-            "parentNameUsageID": np.nan,
-            "acceptedNameUsageID": 5.0,
-            "kingdom": "Animalia",
-            "phylum": "Chordata",
-            "class": "Mammalia",
-            "order": "Carnivora",
-            "family": "Felidae",
-            "genus": "Felis",
-        },
-        {
+        "Puma concolor" : {
             "taxonID": 5,
             "canonicalName": "Puma concolor",
             "taxonomicStatus": "accepted",
@@ -290,23 +71,29 @@ def _create_gbif_tsv_with_taxonomy(tmp_path: Path) -> str:
             "family": "Felidae",
             "genus": "Puma",
         },
-    ]
-    df = pd.DataFrame(rows)
-    fp = tmp_path / "gbif_with_taxonomy.tsv"
-    df.to_csv(fp, sep="\t", index=False)
+    }
+    fp = tmp_path / "gbif_with_taxonomy.json"
+    pd.DataFrame.from_dict(lookupdict, orient="index").to_json(fp, indent=2)
     return str(fp)
 
+def test_gbif_converter() -> None:
+    converter = GBIFConverter()
+    info, ok = converter("Puma concolor")
+    assert ok
+
+    info, ok = converter("Fraudulus animalaticus")
+    assert not ok
 
 def test_add_taxonomy_basic(tmp_path: Path) -> None:
     """Test basic AddTaxonomy transform functionality."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
     # Create test data
     df = pd.DataFrame({"scientific_name": ["Corvus corax", "Passer domesticus"]})
     backend = PandasBackend(df)
 
     # Apply transform
-    transform = AddTaxonomy(feature="scientific_name", gbif_taxonomy_path=gbif_path, use_precomputed_outputs=False)
+    transform = AddTaxonomy(feature="scientific_name", gbif_precomputed_taxonomy_path=gbif_path)
     result_backend, metadata = transform(backend)
 
     # Check that taxonomy columns were added
@@ -327,13 +114,13 @@ def test_add_taxonomy_basic(tmp_path: Path) -> None:
 
 def test_add_taxonomy_from_config(tmp_path: Path) -> None:
     """Test AddTaxonomy.from_config class method."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
     config = AddTaxonomyConfig(
         type="add_taxonomy",
         feature="species",
-        gbif_taxonomy_path=gbif_path,
-        use_precomputed_outputs=False,
+        gbif_precomputed_taxonomy_path=gbif_path,
+
     )
     transform = AddTaxonomy.from_config(config)
 
@@ -351,12 +138,12 @@ def test_add_taxonomy_from_config(tmp_path: Path) -> None:
 
 def test_add_taxonomy_missing_feature_column(tmp_path: Path) -> None:
     """Test that AddTaxonomy raises ValueError for missing feature column."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
     df = pd.DataFrame({"wrong_column": ["Corvus corax"]})
     backend = PandasBackend(df)
 
-    transform = AddTaxonomy(feature="scientific_name", gbif_taxonomy_path=gbif_path, use_precomputed_outputs=False)
+    transform = AddTaxonomy(feature="scientific_name", gbif_precomputed_taxonomy_path=gbif_path)
 
     with pytest.raises(ValueError, match="Feature column 'scientific_name' not found in data"):
         transform(backend)
@@ -364,7 +151,7 @@ def test_add_taxonomy_missing_feature_column(tmp_path: Path) -> None:
 
 def test_add_taxonomy_metadata(tmp_path: Path) -> None:
     """Test that AddTaxonomy returns correct metadata."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
     # Create test data with some duplicates and one unresolvable name
     df = pd.DataFrame(
@@ -379,7 +166,7 @@ def test_add_taxonomy_metadata(tmp_path: Path) -> None:
     )
     backend = PandasBackend(df)
 
-    transform = AddTaxonomy(feature="scientific_name", gbif_taxonomy_path=gbif_path, use_precomputed_outputs=False)
+    transform = AddTaxonomy(feature="scientific_name", gbif_precomputed_taxonomy_path=gbif_path)
     _, metadata = transform(backend)
 
     assert metadata["feature"] == "scientific_name"
@@ -389,12 +176,12 @@ def test_add_taxonomy_metadata(tmp_path: Path) -> None:
 
 def test_add_taxonomy_unresolvable_names(tmp_path: Path) -> None:
     """Test AddTaxonomy with names that cannot be resolved."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
     df = pd.DataFrame({"scientific_name": ["Unknown species", "Another unknown"]})
     backend = PandasBackend(df)
 
-    transform = AddTaxonomy(feature="scientific_name", gbif_taxonomy_path=gbif_path, use_precomputed_outputs=False)
+    transform = AddTaxonomy(feature="scientific_name", gbif_precomputed_taxonomy_path=gbif_path)
     result_backend, metadata = transform(backend)
 
     result_df = result_backend.unwrap
@@ -409,14 +196,14 @@ def test_add_taxonomy_unresolvable_names(tmp_path: Path) -> None:
 
 def test_add_taxonomy_with_add_taxonomic_name(tmp_path: Path) -> None:
     """Test AddTaxonomy with add_taxonomic_name=True."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
     config = AddTaxonomyConfig(
         type="add_taxonomy",
         feature="scientific_name",
-        gbif_taxonomy_path=gbif_path,
+        gbif_precomputed_taxonomy_path=gbif_path,
         add_taxonomic_name=True,
-        use_precomputed_outputs=False,
+
     )
     transform = AddTaxonomy.from_config(config)
 
@@ -432,9 +219,9 @@ def test_add_taxonomy_with_add_taxonomic_name(tmp_path: Path) -> None:
 
 def test_add_taxonomy_make_taxonomic_name(tmp_path: Path) -> None:
     """Test the _make_taxonomic_name method."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
-    transform = AddTaxonomy(feature="scientific_name", gbif_taxonomy_path=gbif_path, use_precomputed_outputs=False)
+    transform = AddTaxonomy(feature="scientific_name", gbif_precomputed_taxonomy_path=gbif_path)
 
     info = {
         "kingdom": "Animalia",
@@ -459,9 +246,9 @@ def test_add_taxonomy_make_taxonomic_name(tmp_path: Path) -> None:
 
 def test_add_taxonomy_make_taxonomic_name_empty(tmp_path: Path) -> None:
     """Test _make_taxonomic_name with empty info."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
-    transform = AddTaxonomy(feature="scientific_name", gbif_taxonomy_path=gbif_path, use_precomputed_outputs=False)
+    transform = AddTaxonomy(feature="scientific_name", gbif_precomputed_taxonomy_path=gbif_path)
 
     result = transform._make_taxonomic_name({})
 
@@ -474,19 +261,19 @@ def test_add_taxonomy_config_validation(tmp_path: Path) -> None:
         AddTaxonomyConfig(
             type="add_taxonomy",
             feature="scientific_name",
-            gbif_taxonomy_path="/nonexistent/path/to/file.tsv",
-            use_precomputed_outputs=False,
+            gbif_precomputed_taxonomy_path="/nonexistent/path/to/file.tsv",
+
         )
 
 
 def test_add_taxonomy_empty_dataframe(tmp_path: Path) -> None:
     """Test AddTaxonomy with an empty dataframe."""
-    gbif_path = _create_gbif_tsv_with_taxonomy(tmp_path)
+    gbif_path = _create_gbif_json_with_taxonomy(tmp_path)
 
     df = pd.DataFrame({"scientific_name": []})
     backend = PandasBackend(df)
 
-    transform = AddTaxonomy(feature="scientific_name", gbif_taxonomy_path=gbif_path, use_precomputed_outputs=False)
+    transform = AddTaxonomy(feature="scientific_name", gbif_precomputed_taxonomy_path=gbif_path)
     result_backend, metadata = transform(backend)
 
     assert len(result_backend) == 0
@@ -511,7 +298,7 @@ def test_add_taxonomy_integration_with_beanszero() -> None:
     transform = AddTaxonomy(
         feature="output",  # 'output' column has the canonical names in BeansZero
         add_taxonomic_name=True,
-        use_precomputed_outputs=False,
+
         # Uses cache if present
     )
 
@@ -543,51 +330,3 @@ def test_add_taxonomy_integration_with_beanszero() -> None:
         taxonomic_name = row["taxonomic_name"]
         assert taxonomic_name is not None
         assert "Animalia" in taxonomic_name
-
-def test_precomputed_vs_tsv_call_paths_consistent_on_10_lookupdict_keys(tmp_path: Path) -> None:
-    """
-    Consistency of results across the two __call__ pathways:
-
-    - Build precomputed converter (fast path) using default gs:// JSON, and take 10 keys from lookupdict.keys().
-    - Build TSV-backed converter (slow path) using default gs:// TSV.
-    - For each of the 10 keys, ensure __call__ returns identical (info, ok) tuples.
-    """
-
-    # Put caches in tmp so CI doesn't attempt to write into the package directory.
-    tsv_cache = tmp_path / "gbif_animals_cache.tsv"
-    json_cache = tmp_path / "gbif_animals_converter_cache.json"
-
-    c_pre = GBIFConverter(
-        use_precomputed_outputs=True,
-        precomputed_cache_path=str(json_cache),
-    )
-
-    assert c_pre.lookupdict is not None
-    keys = sorted(c_pre.lookupdict.keys())[:10]
-    assert len(keys) == 10, "Expected at least 10 keys in lookupdict from precomputed cache"
-
-    c_tsv = GBIFConverter(
-        use_precomputed_outputs=False,
-        cache_path=str(tsv_cache),
-    )
-
-    def _normalize_missing(d: dict) -> dict:
-        """
-        convert nan values to None
-
-        Returns
-        --------
-        Dict with modified values
-        """
-        return {
-            k: (None if pd.isna(v) else v)
-            for k, v in d.items()
-        }
-
-    for name in keys:
-        info_pre, ok_pre = c_pre(name)
-        info_tsv, ok_tsv = c_tsv(name)
-
-        assert ok_pre is True
-        assert ok_tsv is True
-        assert _normalize_missing(info_pre) == _normalize_missing(info_tsv)
